backend/pyproject.toml
[project]
name = "ai-game-platform-backend"
version = "0.1.0"
description = "FastAPI backend for AI Game Platform"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.1",
  "redis>=5.0.7",
  "PyYAML>=6.0.2",
  "PyJWT>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
  "httpx>=0.27.0",
  "ruff>=0.6.2",
  "mypy>=1.11.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
